+++ src/Domain/Campaign/Actions/PrepareEmailHtmlAction.php
@@ -53,16 +53,15 @@ class PrepareEmailHtmlAction
 
     private function addUtmTags(Campaign $campaign): void
     {
-        $replacements = $campaign->htmlLinks()
-            ->mapWithKeys(function (string $link) use ($campaign) {
-                $newLink = $this->addUtmTagsToUrlAction->execute($link, $campaign->name);
+        $document = $this->createDomDocumentFromHtmlAction->execute($campaign->email_html);
 
-                return [$link => $newLink];
-            });
+        /** @var \DOMElement $linkElement */
+        foreach ($document->getElementsByTagName('a') as $linkElement) {
+            $url = $linkElement->getAttribute('href');
+            $newUrl = $this->addUtmTagsToUrlAction->execute($url, $campaign->name);
+            $linkElement->setAttribute('href', $newUrl);
+        }
 
-        $campaign->email_html = strtr(
-            $campaign->email_html,
-            $replacements->toArray(),
-        );
+        $campaign->email_html = $document->saveHTML();
     }
 }