+++ src/Http/App/Queries/Filters/CampaignStatusFilter.php
@@ -21,10 +21,6 @@ class CampaignStatusFilter implements Filter
             return $query->draft();
         }
 
-        if ($value === 'automated') {
-            return $query->automated();
-        }
-
         return $query;
     }
 }